Undo/Redo accessibility & multiplayer bug fix
multi-player bug fix is a damn annoying one.
in some scenarios, it will allow players to push each other, instead of trying to propogate from them moving at the same time and out of each others way.
thats an edge case, but exactly the kind of edge-case that my system here should be able to prevent.
but it doesnt because the entire multi-player system is a hack on the grid framework, and really shouldnt be used tbh.
(except i need to fix that because AI moving at the same time?)